[SECURITY] Verify LZW code fits in 12 bits before we use it
The type `unsigned short`, which is the type of the LZW code, is much
larger than the max of an LZW code, causing a wild pointer.

That causes an array overflow. Long term, this library should jettison
this unnecessary GIF code, and use a better maintained library for it.
In the meantime, however, the issue is simple enough to solve, so I've
done so.

Resolves CVE-2020-19668.
Closes saitoha/libsixel#136.
Closes #7.