sixel_encoder_prepare_palette: kill use-after-free
The only place sixel_encoder_prepare_palette() is called
from is sixel_encoder_encode_frame(). If we have a
SIXEL_FAILED result inside the former, we unref the palette,
potentially freeing it. We then goto the error path for
sixel_encoder_encode_frame(), where we unref it again, in a
classic double-free. Remove the internal unref. As noted,
this is a single call site, so removing the unref can't
cause a memory leak.

Closes #27, reported by a4865g ("WuLearn").